backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langgraph==0.2.36
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
pdfplumber==0.11.4
python-docx==1.1.2
python-multipart==0.0.9
groq>=0.11.0
